+++ sf-review-application/src/main/java/com/hexalyte/sf_review_application/service/impl/ReviewServiceImpl.java
@@ -1,4 +1,4 @@
-package com.hexalyte.sf_review_application.service.impl; // Note: package might be service.impl or just service depending on your structure, check existing file.
+package com.hexalyte.sf_review_application.service.impl;
 
 import com.hexalyte.sf_review_application.model.Review;
 import com.hexalyte.sf_review_application.repository.ReviewRepository;
@@ -7,7 +7,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
-import java.util.UUID; // Import UUID
+import java.util.UUID; // FIX: Import UUID
 
 @Service
 public class ReviewServiceImpl implements ReviewService {
@@ -26,12 +26,13 @@ public class ReviewServiceImpl implements ReviewService {
                 .orElseThrow(() -> new RuntimeException("Review not found with id: " + reviewId));
     }
 
+    // FIX: Changed to Long
     @Override
-    public List<Review> getReviewsByServiceProvider(Integer serviceProviderId) {
+    public List<Review> getReviewsByServiceProvider(Long serviceProviderId) {
         return reviewRepository.findByServiceProviderId(serviceProviderId);
     }
 
-    // FIX: Change parameter type to UUID
+    // FIX: Changed to UUID
     @Override
     public List<Review> getReviewsByUser(UUID userId) {
         return reviewRepository.findByUserId(userId);
@@ -45,8 +46,9 @@ public class ReviewServiceImpl implements ReviewService {
         reviewRepository.deleteById(reviewId);
     }
 
+    // FIX: Changed to Long
     @Override
-    public Double getAverageRating(Integer serviceProviderId) {
+    public Double getAverageRating(Long serviceProviderId) {
         List<Review> reviews = reviewRepository.findByServiceProviderId(serviceProviderId);
         if (reviews.isEmpty()) {
             return 0.0;
